Most of the project
Project has intertwined PDO, Fetch and MySqli. Used all 3 for learning purposes. I'm aware it's against the best practices of the field.

Also a lot of commenting and translating comments would still be required.

Also need to clean old parts of HTML and some PHP that are not in use.